Working on the new rest node
